+++ src/components/App/App.scss
@@ -3,7 +3,3 @@
 .App {
   font-family: font-family(heading);
 }
-
-.Loader {
-  text-align: center;
-}
