+++ src/components/ConferencePage/ConferencePage.scss
@@ -1,0 +1,3 @@
+.Loader {
+  text-align: center;
+}
